+++ muduo/base/tests/AsyncLogging_test.cc
@@ -4,8 +4,9 @@
 
 #include <stdio.h>
 #include <sys/resource.h>
+#include <unistd.h>
 
-int kRollSize = 500*1000*1000;
+off_t kRollSize = 500*1000*1000;
 
 muduo::AsyncLogging* g_asyncLog = NULL;
